+++ apps/wallet/src/ui/app/pages/home/tokens/TokenLink.tsx
@@ -1,7 +1,8 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-import { MIST_PER_SUI, type CoinBalance } from '@mysten/sui.js';
+import { type CoinBalance } from '@mysten/sui.js/client';
+import { MIST_PER_SUI } from '@mysten/sui.js/utils';
 import { type ReactNode } from 'react';
 import { Link } from 'react-router-dom';