+++ crates/sui-graphql-rpc/src/context_data/mod.rs
@@ -1,0 +1,6 @@
+// Copyright (c) Mysten Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+pub(crate) mod context_ext;
+pub(crate) mod data_provider;
+pub(crate) mod sui_sdk_data_provider;