+++ crates/sui-graphql-rpc/src/types/object.rs
@@ -11,8 +11,9 @@ use super::{
     balance::Balance, coin::Coin, owner::Owner, stake::Stake, sui_address::SuiAddress,
     transaction_block::TransactionBlock,
 };
-use crate::server::sui_sdk_data_provider::SuiClientLoader;
-use crate::{server::context_ext::DataProviderContextExt, types::base64::Base64};
+use crate::context_data::context_ext::DataProviderContextExt;
+use crate::context_data::sui_sdk_data_provider::SuiClientLoader;
+use crate::types::base64::Base64;
 
 #[derive(Clone, Eq, PartialEq, Debug)]
 pub(crate) struct Object {