+++ sdk/dapp-kit/src/hooks/useSuiClientMutation.ts
@@ -7,17 +7,15 @@ import { useSuiClientContext } from './useSuiClient.js';
 import type { SuiRpcMethods } from './useSuiClientQuery.js';
 
 export type UseSuiClientMutationOptions<T extends keyof SuiRpcMethods> = Omit<
-	UseMutationOptions<SuiRpcMethods[T]['result'], Error, SuiRpcMethods[T]['result'], unknown[]>,
+	UseMutationOptions<SuiRpcMethods[T]['result'], Error, SuiRpcMethods[T]['params'], unknown[]>,
 	'mutationFn'
 >;
 
 export function useSuiClientMutation<T extends keyof SuiRpcMethods>(
 	{
 		method,
-		params,
 	}: {
 		method: T;
-		params: SuiRpcMethods[T]['params'];
 	},
 	options: UseSuiClientMutationOptions<T> = {},
 ) {
@@ -25,7 +23,7 @@ export function useSuiClientMutation<T extends keyof SuiRpcMethods>(
 
 	return useMutation({
 		...options,
-		mutationFn: async () => {
+		mutationFn: async (params) => {
 			return await suiContext.client[method](params as never);
 		},
 	});