+++ sdk/dapp-kit/src/index.ts
@@ -12,3 +12,4 @@ export * from './hooks/wallet/useWallet.js';
 export * from './hooks/useSuiClientMutation.js';
 export * from './hooks/useSuiClientQuery.js';
 export * from './hooks/useSuiClientInfiniteQuery.js';
+export * from './types.js';